requirements.txt
python-nmap
termcolor
shodan
pexpect
tqdm
